+++ src/main/java/com/Moleugo/moleugo/entity/UserActivity.java
@@ -1,0 +1,4 @@
+package com.Moleugo.moleugo.entity;
+
+public class UserActivity {
+}
