+++ src/main/java/com/Moleugo/moleugo/service/member/auth/FindPasswordService.java
@@ -1,0 +1,4 @@
+package com.Moleugo.moleugo.service.member.auth;
+
+public class FindPasswordService {
+}
