+++ src/main/java/com/Moleugo/moleugo/service/member/auth/LoginService.java
@@ -1,9 +1,9 @@
-package com.Moleugo.moleugo.service.member;
+package com.Moleugo.moleugo.service.member.auth;
 
 import com.Moleugo.moleugo.entity.Member;
 import com.Moleugo.moleugo.repository.MemberRepository;
 import com.Moleugo.moleugo.response.LoginResponse;
-import com.Moleugo.moleugo.service.validator.LoginValidator;
+import com.Moleugo.moleugo.validator.LoginValidator;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
