+++ src/main/java/com/Moleugo/moleugo/validator/NicknameValidator.java
@@ -1,4 +1,4 @@
-package com.Moleugo.moleugo.service.validator;
+package com.Moleugo.moleugo.validator;
 
 import com.Moleugo.moleugo.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
